+++ aws/modules/eks/outputs.tf
@@ -1,0 +1,13 @@
+output "eks_cluster_iam_role" {
+  value = {
+    name = aws_iam_role.eks_role.name,
+    arn  = aws_iam_role.eks_role.arn
+  }
+}
+
+output "eks_node_iam_role" {
+  value = {
+    name = aws_iam_role.node_group_role.name
+    arn  = aws_iam_role.node_group_role.arn
+  }
+}