+++ aws/modules/vpc/main.tf
@@ -15,7 +15,9 @@ resource "aws_subnet" "ts_aws_subnet_public_igw" {
   vpc_id            = aws_vpc.ts_aws_vpc.id
   cidr_block        = cidrsubnet(aws_vpc.ts_aws_vpc.cidr_block, var.ts_aws_vpc_cidr_newbits, each.value)
   availability_zone = each.key
-
+  tags = {
+    "kubernetes.io/role/elb" = 1
+  }
   map_public_ip_on_launch = true
 }
 
@@ -27,7 +29,7 @@ resource "aws_route_table" "ts_aws_route_table_public_igw" {
 
 resource "aws_route_table_association" "ts_aws_route_table_association_public_igw" {
 
-  for_each       = aws_subnet.ts_aws_subnet_public_igw
+  for_each = aws_subnet.ts_aws_subnet_public_igw
 
   subnet_id      = each.value.id
   route_table_id = aws_route_table.ts_aws_route_table_public_igw.id
@@ -85,7 +87,7 @@ resource "aws_nat_gateway" "ts_aws_nat_gateway" {
   subnet_id     = aws_subnet.ts_aws_subnet_public_igw[each.key].id
   allocation_id = aws_eip.ts_aws_eip_nat[each.key].id
 
-  depends_on    = [aws_internet_gateway.ts_aws_internet_gateway]
+  depends_on = [aws_internet_gateway.ts_aws_internet_gateway]
 }
 
 resource "aws_route" "ts_aws_route_private_ngw" {