+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 0.3.3
+----
+Add EKS module for AWS.
+Modified aws alb module to allow creation of only the alb.
+Added an eks_alb_controller module to use with the EKS cluster.
+Added local helm template example.
+
+0.3.3
 ---
 Added standards doc